+++ .env
@@ -1,2 +1,7 @@
-MONGO_URI=mongodb://localhost:27017/ibm-ride-app
+MONGO_URI=mongodb://localhost:27017/ride-app  
+PORT=3000
+DB_HOST=localhost
+DB_USER=root
+DB_PASS=Laxmikanth123*
+DB_NAME=testdb
 PORT=3000
