+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "dotenv": "^17.2.2",
     "express": "^5.1.0",
-    "mongoose": "^8.18.0"
+    "mongoose": "^8.18.0",
+    "mysql2": "^3.14.4"
   }
 }
